+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+### Unreleased
+
+### Changed
+
+- The `DateObserved` alert query filter now parses timestamps using microsecond precision.
+
+### Fixed
+
+- Python2.7 bug where non-unicode `str` timestamps would fail when used in a query filter.
+
 ## 1.15.0 - 2021-06-16
 
 ### Fixed